feat(create): give /create its own builder top bar
